apps/ai-service/requirements.txt
# Web framework
fastapi==0.135.3
uvicorn[standard]==0.44.0
python-multipart==0.0.24

# OpenAI
openai==2.31.0

# LangGraph + LangChain
langgraph==1.1.6
langgraph-prebuilt==1.0.9
langchain==1.2.15
langchain-openai==1.1.12
langchain-core==1.2.28
langchain-community==0.3.31
langchain-text-splitters==0.3.11

# Supabase
supabase==2.28.3

# Neo4j
neo4j==5.28.3

# Vector DB
qdrant-client==1.17.1
fastembed==0.8.0

# Bi-temporal knowledge graph
graphiti-core==0.28.2

# Multi-scope memory
mem0ai==1.0.11

# Guardrails
nemoguardrails==0.21.0        # L2 dialog rails
presidio-analyzer==2.2.362    # L1/L3 PHI/PII scan (Python 3.13 compatible)
presidio-anonymizer==2.2.362

# Observability
langfuse==4.1.0

# PDF
pypdf==4.3.1

# Utilities
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-jose[cryptography]==3.3.0
tiktoken==0.12.0
tenacity==9.0.0
structlog==25.5.0
httpx==0.28.1
